+++ scripts/codesign.py
@@ -1,9 +1,17 @@
 #!/usr/bin/env python3
-
+# This script assumes/requires a valid codesign certificate.
+# To add a codesign certificate programmatically:
+# 1. Unlock the login keychain:
+#   security unlock-keychain login.keychain
+# 2. Import the codesign ("Developer ID Application") certificate into the login keychain:
+#   security import <CERTIFICATE.p12> -k ~/Library/Keychains/login.keychain-db -P <PASSWORD> -T /usr/bin/codesign
+# 3. Set the ACL on the keychain:
+#   security set-key-partition-list -S apple-tool:,apple: -s -k <PASSWORD> [KEYCHAIN]
+# Source/reference: https://stackoverflow.com/a/52115968
+import sys
 from configparser import RawConfigParser
 from pathlib import Path
 from subprocess import run
-import sys
 
 
 def codesign_app(developer_id: str, path: str) -> None:
@@ -20,6 +28,7 @@ def codesign_app(developer_id: str, path: str) -> None:
         ]
     )
 
+
 def codesign_dmg(developer_id: str, path: str) -> None:
     run(
         [
@@ -32,11 +41,14 @@ def codesign_dmg(developer_id: str, path: str) -> None:
         ]
     )
 
+
 def codesign_verify(path: str):
     run(["codesign", "--verify", "--verbose=1", path])
 
+
 def codesign_display(path: str):
     run(["codesign", "--display", "--verbose=4", path])
+
 
 def spctl_assess_app(path: str):
     run(
@@ -48,6 +60,7 @@ def spctl_assess_app(path: str):
             path,
         ]
     )
+
 
 def spctl_assess_dmg(path: str):
     run(